Factored out common code into util.py.
Important refactoring affecting most of the python scripts. Shared functions and
global constants are now in util.py

The command line interface has changed for most of the scripts. By default,
the path of generated files doesn't have to be explicitly given anymore.

A last change is the serialization of the global weights for tags, generated in
genmat.py. These weights are needed when new / unseen tracks need to be
projected in the latent space.