Compute similarity from JSON files.
This diff changes the semantics of 'similarity.py'. Until now, the tracks were
specified using a minimalist pseudo-language directly on the command line. From
now on, it compares tracks based on a JSON file containig their metadata.

The file should have the format of the MSD's last.fm dataset. Basically:

    {'artist': ".", 'title': ".", 'tags': [['rock', 1.0], ['pop', 0.0]]}