Slight changes to JSON output structure.
Improves the consistency across the the whole system. A room is systematically
referred to by its id, `rid`. Same for a user (`uid`). Password is not
abbreviated `pass` anymore -- use `password` instead.

Also, a call to the API root now returns the user's nickname and the room he's
supposed to be in, in addition to his uid. This spares us two calls on login in
the android application.