Basic acknowledgment for the job queue.
I don't seem to be able to read messages from the RabbitMQ queue one after the
other. This means that we lose a lot of messages if a big batch of them comes at
once and fetcher.py fails (which happens e.g. when a new user upload his whole
library).

This diff mitigates this problem by sending an acknowledgment only once we start
processing a job.